abstract initialization
